Move data_generators.py to train_models folder
