+++ ms2deepscore/train_new_model/train_ms2deepscore.py
@@ -5,8 +5,8 @@ This script is not needed for normally running MS2Deepscore, it is only needed t
 import os
 from typing import Optional
 from matplotlib import pyplot as plt
-from ms2deepscore.data_generators import (DataGeneratorPytorch,
-                                          TensorizationSettings)
+from ms2deepscore.train_new_model.data_generators import (DataGeneratorPytorch,
+                                                          TensorizationSettings)
 from ms2deepscore.models.SiameseSpectralModel import (SiameseSpectralModel,
                                                       train)
 from ms2deepscore.SettingsMS2Deepscore import (GeneratorSettings,
